Move metaverse-modules into core-modules
