+++ core-modules.js
