+++ metaversefile-api.js
@@ -26,7 +26,7 @@ import {world} from './world.js';
 // import ERC721 from './erc721-abi.json';
 // import ERC1155 from './erc1155-abi.json';
 // import {web3} from './blockchain.js';
-import {moduleUrls, importModule} from './metaverse-modules.js';
+import {moduleUrls, importModule} from './core-modules.js';
 import {componentTemplates} from './metaverse-components.js';
 import postProcessing from './post-processing.js';
 import {getRandomString, memoize} from './util.js';
