+++ webaverse.js
@@ -34,7 +34,6 @@ import {
   getComposer,
 } from './renderer.js';
 import transformControls from './transform-controls.js';
-// import * as metaverseModules from './metaverse-modules.js';
 import dioramaManager from './diorama.js';
 import * as voices from './voices.js';
 import performanceTracker from './performance-tracker.js';
